backend/pyproject.toml
[project]
name = "prelegal-backend"
version = "0.1.0"
description = "Prelegal API backend"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.34.0",
    "python-dotenv>=1.0.0",
    "pydantic>=2.0.0",
    "pydantic-settings>=2.0.0",
    "structlog>=24.0.0",
    "markdown2>=2.5.0",
    "weasyprint>=63.0",
    "jinja2>=3.1.0",
    "python-jose[cryptography]>=3.5.0",
    "aiosqlite>=0.22.1",
    "email-validator>=2.3.0",
    "bcrypt>=5.0.0",
]

[dependency-groups]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.24.0",
    "httpx>=0.27.0",
    "ruff>=0.8.0",
    "mypy>=1.13.0",
    "pytest-cov>=7.0.0",
]

[tool.ruff]
line-length = 88
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "I"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[tool.coverage.run]
source = ["app"]
omit = ["app/__init__.py"]

[tool.mypy]
python_version = "3.12"
strict = true
